Ensured averages will be accurate.

+++ Datalogger/Datalogger.ino
@@ -148,10 +148,10 @@ String takeMeasurement(String setting) {
   }
   Serial.println("");
   Serial.println("Finished getting values. Averaging...");
-  averageValue1 = sumValuesSensor1/30; //And this divides it by 30 (it takes one sample per second, for 30 seconds)
-  averageValue2 = sumValuesSensor2/30;
-  averageValue3 = sumValuesSensor3/30;
-  averageValue4 = sumValuesSensor4/30;
+  averageValue1 = ((double)sumValuesSensor1)/30; //And this divides it by 30 (it takes one sample per second, for 30 seconds)
+  averageValue2 = ((double)sumValuesSensor2)/30;
+  averageValue3 = ((double)sumValuesSensor3)/30;
+  averageValue4 = ((double)sumValuesSensor4)/30;
   Serial.println("Putting data into dataString");
   data += "averageValue1";
   data += ",";
